WIP: handle 'launcher-additional-settings' for module search.
The idea here was to add 'launcherAdditionalSettings' to correspond to the existing 'userRevisionSettings'.
But I don't want to have one settings file where we only look at '[Modules/AdditionalPaths]' --
that seems confusing and error-prone because users will expect it to work generally.
Need a better way to do it. For now will handle in CMake.

ref: http://www.na-mic.org/Bug/view.php?id=4242