ENH: add multishell DWI test dataset
Small patch of multi-shell phantom dataset from Tristan Kuder.
Used with permission. No PHI issues because it is a phantom.

The volume was reduced to a few voxels, to minimize file size and
test time. As a result, this dataset is probably only useful for i/o
tests. We gratefully thank Dr. Kuder, and in the event of any
research use, please cite the papers mentioned in the following
correspondence:

```
Isaiah Norton:
>> Would it be acceptable to redistribute this DICOM data publicly, under
>> something like PDDL [1] or a similar license? I am interested in adding
>> (a subset of) this data to the public regression test cases for the
>> "3D Slicer" diffusion DICOM converter.

Tristan Kuder:
>  We are glad that you find the dataset useful. Please feel free to use
>  it for the purpose you mentioned in your email.

> If possible, an acknowledgement for Dr. Frederik Laun and myself would
> be nice and especially in this context an acknowledgement for the grant
> numbers DFG LA 2804/1-1 and LA 2804/1-2.

> Perhaps, you could also cite the initial paper regarding the phantoms:
> http://www.mrijournal.com/article/S0730-725X%2808%2900292-0/abstract

> You could also cite the subsequent papers:
> http://onlinelibrary.wiley.com/doi/10.1002/mrm.22602/abstract
> http://onlinelibrary.wiley.com/doi/10.1002/mrm.24774/abstract